+++ test/sql/functions.sql
@@ -35,6 +35,7 @@ SELECT clickhouse_raw_query('CREATE TABLE functions_test.t4 (val String) engine=
 SELECT clickhouse_raw_query('CREATE TABLE functions_test.t5 (ts DateTime) engine=TinyLog();');
 SELECT clickhouse_raw_query('CREATE TABLE functions_test.t6 (i64 Int64, f64 Float64) engine=TinyLog();');
 SELECT clickhouse_raw_query('CREATE TABLE functions_test.t7(dt Date) engine=TinyLog();');
+SELECT clickhouse_raw_query('CREATE TABLE functions_test.t8(val String) engine=TinyLog();');
 
 SELECT clickhouse_raw_query($$
 	INSERT INTO functions_test.t5 VALUES
@@ -72,6 +73,7 @@ CREATE FOREIGN TABLE t4 (val text) SERVER functions_loopback;
 CREATE FOREIGN TABLE t5 (ts timestamp) SERVER functions_loopback;
 CREATE FOREIGN TABLE t6 (i64 BIGINT, f64 FLOAT8) SERVER functions_loopback;
 CREATE FOREIGN TABLE t7 (ts date) SERVER functions_loopback;
+CREATE FOREIGN TABLE t8 (val text) SERVER functions_loopback;
 
 SELECT clickhouse_raw_query($$
 	INSERT INTO functions_test.t3
@@ -89,6 +91,14 @@ SELECT clickhouse_raw_query($$
 	INSERT INTO functions_test.t4
 	SELECT 'val' || toString(number+1)
 	  FROM numbers(2);
+$$);
+
+SELECT clickhouse_raw_query($$
+	INSERT INTO functions_test.t8
+	VALUES ('a,b,c'),
+		   ('foo,bar,baz'),
+		   ('sleep   no   more'),
+		   ('aa-T-bb-t-cc')
 $$);
 
 SELECT clickhouse_raw_query($$
@@ -306,8 +316,18 @@ EXPLAIN (VERBOSE, COSTS OFF) SELECT ts FROM t5 WHERE date_trunc('quarter', ts) =
 SELECT ts FROM t5 WHERE date_trunc('quarter', ts) = '2027-10-01'::date;
 
 -- check regexp_like.
-EXPLAIN (VERBOSE, COSTS OFF) SELECT val FROM t4 WHERE regexp_like('^val\d', val);
-SELECT val FROM t4 WHERE regexp_like('^val\d', val);
+EXPLAIN (VERBOSE, COSTS OFF) SELECT val FROM t4 WHERE regexp_like(val, '^val\d');
+SELECT val FROM t4 WHERE regexp_like(val, '^val\d');
+EXPLAIN (VERBOSE, COSTS OFF) SELECT val FROM t4 WHERE regexp_like(val, '^VAL\d', 'i');
+SELECT val FROM t4 WHERE regexp_like(val, '^VAL\d', 'i');
+
+-- Check regexp_split_to_array.
+EXPLAIN (VERBOSE, COSTS OFF) SELECT val FROM t8 WHERE regexp_split_to_array(val, ',') = '{a,b,c}'::text[];
+SELECT val FROM t8 WHERE regexp_split_to_array(val, ',') = '{a,b,c}'::text[];
+EXPLAIN (VERBOSE, COSTS OFF) SELECT val FROM t8 WHERE regexp_split_to_array(val, '\s+') = '{sleep,no,more}'::text[];
+SELECT val FROM t8 WHERE regexp_split_to_array(val, '\s+') = '{sleep,no,more}'::text[];
+EXPLAIN (VERBOSE, COSTS OFF) SELECT val FROM t8 WHERE regexp_split_to_array(val, '-t-', 'i') = '{aa,bb,cc}'::text[];
+SELECT val FROM t8 WHERE regexp_split_to_array(val, '-t-', 'i') = '{aa,bb,cc}'::text[];
 
 -- Check hashing functions.
 EXPLAIN (VERBOSE, COSTS OFF) SELECT key1, val FROM t3_map WHERE md5(val) LIKE 'a%' ORDER BY key1;